Onboarding: Rename CV settings screen
